+++ backend/src/db.ts
@@ -5,7 +5,7 @@ import "dotenv/config";
 
 const databaseFileName = `file:${path.join(
   process.env.DB_FILE_DIR ?? "../database",
-  process.env.DB_FILE_NAME ?? "database.db",
+  process.env.DB_FILE_NAME ?? "database.db"
 )}`;
 
 export const client = createClient({
