+++ backend/src/routes/index.ts
@@ -1,6 +1,9 @@
 import type { FastifyInstance } from "fastify";
-import usersRoutes from "./users";
+import usersRoutes from "./user/user";
+
+import gameIdRoute from "./game/socket";
 
 export default async function routes(fastify: FastifyInstance) {
-  fastify.register(usersRoutes, { prefix: "/api/users" });
+  fastify.register(usersRoutes, { prefix: "/user" });
+  fastify.register(gameIdRoute, { prefix: "/game" });
 }
